+++ packages/playground/src/main.ts
@@ -1,5 +1,6 @@
 import { createApp } from 'vue'
 import { App as VineApp } from './vine/app.vine'
+import router from './router'
 
 // UnoCSS
 import 'virtual:uno.css'
@@ -7,4 +8,5 @@ import 'virtual:uno.css'
 import './styles/main.css'
 
 const app = createApp(VineApp)
+app.use(router)
 app.mount('#app')
